requirements.txt
fastapi==0.128.0
langchain==1.2.7
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-ollama==1.0.1
langchain-openai==1.1.7
langgraph==1.0.7
openai==2.24.0
pydantic[email]==2.12.5
pydantic-settings==2.12.0
streamingjson==0.0.5
uvicorn==0.41.0
langgraph-checkpoint-mongodb==0.3.1
python-dotenv == 1.2.2